Added and refactored to use set-text-from-dom-node with improved set selection after input
